data_scientist_agent/src/agent.py
# src/agent.py

import re
from dotenv import load_dotenv
from langchain_google_genai import ChatGoogleGenerativeAI

# The core prompt that instructs the LLM to act as a code-generating data scientist
CODE_GENERATOR_PROMPT = """
You are a world-class data scientist and a master of Python. 
Your role is to assist a user by writing Python code to analyze a pandas DataFrame.

You have access to a pandas DataFrame named `df`.

The user will ask a question in plain English. You must respond with a single block of Python code to answer the question.

RULES:
1.  **Code Only:** Your entire response must be ONLY the Python code, wrapped in a single markdown block (e.g., ```python...```). Do not provide any text, explanation, or conversation outside the code block.
2.  **Display Output:** The code you write MUST produce a visible output. 
    - For DataFrames or Series, use `print()` or have it as the last line.
    - For plots, the last line of your code must be the figure object (e.g., `fig`). The environment can display Matplotlib and Plotly figures. Use libraries like `plotly.express` as `px` or `matplotlib.pyplot` as `plt`.
    - For any other value (like a number or string), use `print()`.
3.  **Stateful Modifications:** The `df` object is persistent. Code that modifies it (e.g., `df.dropna(inplace=True)`) will alter the DataFrame for subsequent turns.
4.  **Imports:** All necessary libraries (`pandas as pd`, `numpy as np`, `plotly.express as px`, `matplotlib.pyplot as plt`) are already available. Do not write `import` statements.

Example user question: "Show the first 5 rows of the data"
Your response:
```python
print(df.head())
"""


def get_llm():
    """Initializes and returns the LLM for code generation."""
    load_dotenv()
    return ChatGoogleGenerativeAI(
        model="gemini-1.5-flash",
        temperature=0,
    )


def extract_python_code(text: str) -> str | None:
    """Extracts Python code from a markdown code block."""
    match = re.search(r"python\n(.*?)\n", text, re.DOTALL)
    if match:
        return match.group(1)

    return None
