data_scientist_agent/requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
python-dotenv
scikit-learn
langchain-groq